refactor: drop `user` types file
