+++ src/definitions.ts
@@ -1,10 +1,27 @@
 /// <reference types="@capacitor/cli" />
-import { Authentication, User } from './user';
 
 declare module '@capacitor/cli' {
   export interface PluginsConfig {
     GoogleAuth: GoogleAuthPluginOptions;
   }
+}
+
+export interface User {
+  id: string;
+  email: string;
+
+  name: string;
+  familyName: string;
+  givenName: string;
+  imageUrl: string;
+
+  serverAuthCode: string;
+  authentication: Authentication;
+}
+
+export interface Authentication {
+  accessToken: string;
+  idToken: string;
 }
 
 export interface GoogleAuthPluginOptions {
